drm: Workaround COLOR_ENCODING and COLOR_RANGE pushing us out of planes
We only use YCbCr on primary planes, which is lucky because AMDGPU only supports COLOR_ENCODING and COLOR_RANGE on primary planes.

This may not be the fully correct way to do this liftoff wise, but given our lowest zpos layer is always the base plane due to how we set it up this works for now.